autogen_intelligent_feedback_analysis_action/requirements.txt
# AutoGen-based requirements with compatible versions for Python 3.13

# Core AutoGen
pyautogen>=0.10.0
autogen-agentchat>=0.7.0

# OpenAI
openai>=1.40.0

# Data processing (compatible with Python 3.13)
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0

# NLP
nltk>=3.8.0
textblob>=0.17.1

# Utilities
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0

# Visualization
streamlit>=1.29.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Async
aiohttp>=3.8.0

# Additional
docker>=6.0.0
jupyter>=1.0.0
